+++ latch/__init__.py
@@ -18,8 +18,8 @@ from latch.resources.conditional import create_conditional_section
 from latch.resources.map_tasks import map_task
 from latch.resources.reference_workflow import workflow_reference
 from latch.resources.tasks import (
-    custom_task,
     custom_memory_optimized_task,
+    custom_task,
     large_gpu_task,
     large_task,
     medium_task,