+++ latch/types/file.py
@@ -126,6 +126,9 @@ class LatchFile(FlyteFile):
     def __hash__(self) -> int:
         return hash(self.path)
 
+    def size(self):
+        return LPath(self.remote_path).size()
+
     def _idempotent_set_path(self, hint: Optional[str] = None):
         if self._path_generated:
             return