+++ latch_cli/main.py
@@ -36,7 +36,7 @@ T = TypeVar("T")
 
 
 def requires_login(f: Callable[P, T]) -> Callable[P, T]:
-    def decorated(*args, **kwargs):
+    def decorated(*args: P.args, **kwargs: P.kwargs):
         try:
             get_auth_header()
         except AuthenticationError as e:
@@ -54,6 +54,8 @@ def requires_login(f: Callable[P, T]) -> Callable[P, T]:
             raise click.exceptions.Exit(1) from e
 
         return f(*args, **kwargs)
+
+    decorated.__doc__ = f.__doc__
 
     return decorated
 
@@ -106,8 +108,8 @@ def login(connection: Optional[str]):
     click.secho("Successfully logged into LatchBio.", fg="green")
 
 
-@requires_login
 @main.command("workspace")
+@requires_login
 def workspace():
     """Spawns an interactive terminal prompt allowing users to choose what workspace they want to work in."""
 
@@ -315,7 +317,6 @@ def generate_metadata(
         raise click.exceptions.Exit(1)
 
 
-@requires_login
 @main.command("develop")
 @click.argument("pkg_root", nargs=1, type=click.Path(exists=True, path_type=Path))
 @click.option(
@@ -340,6 +341,7 @@ def generate_metadata(
     type=bool,
     help="Start a develop session for a Snakemake workflow.",
 )
+@requires_login
 def local_development(
     pkg_root: Path,
     yes: bool,
@@ -369,7 +371,6 @@ def local_development(
         local_development(pkg_root.resolve(), snakemake)
 
 
-@requires_login
 @main.command("exec")
 @click.option(
     "--execution-id", "-e", type=str, help="Optional execution ID to inspect."
@@ -381,6 +382,7 @@ def local_development(
     type=int,
     help="Optional container index to inspect (only used for Map Tasks)",
 )
+@requires_login
 def execute(
     execution_id: Optional[str], egn_id: Optional[str], container_index: Optional[int]
 ):
@@ -391,7 +393,6 @@ def execute(
     exec(execution_id=execution_id, egn_id=egn_id, container_index=container_index)
 
 
-@requires_login
 @main.command("register")
 @click.argument("pkg_root", type=click.Path(exists=True, file_okay=False))
 @click.option(
@@ -474,6 +475,7 @@ def execute(
     default=None,
     help="Set execution profile for Nextflow workflow",
 )
+@requires_login
 def register(
     pkg_root: str,
     disable_auto_version: bool,
@@ -516,7 +518,6 @@ def register(
     )
 
 
-@requires_login
 @main.command("launch")
 @click.argument("params_file", nargs=1, type=click.Path(exists=True))
 @click.option(
@@ -524,6 +525,7 @@ def register(
     default=None,
     help="The version of the workflow to launch. Defaults to latest.",
 )
+@requires_login
 def launch(params_file: Path, version: Union[str, None] = None):
     """Launch a workflow using a python parameter map."""
 
@@ -542,7 +544,6 @@ def launch(params_file: Path, version: Union[str, None] = None):
     )
 
 
-@requires_login
 @main.command("get-params")
 @click.argument("wf_name", nargs=1)
 @click.option(
@@ -550,6 +551,7 @@ def launch(params_file: Path, version: Union[str, None] = None):
     default=None,
     help="The version of the workflow. Defaults to latest.",
 )
+@requires_login
 def get_params(wf_name: Union[str, None], version: Union[str, None] = None):
     """Generate a python parameter map for a workflow."""
     crash_handler.message = "Unable to generate param map for workflow"
@@ -567,13 +569,13 @@ def get_params(wf_name: Union[str, None], version: Union[str, None] = None):
     )
 
 
-@requires_login
 @main.command("get-wf")
 @click.option(
     "--name",
     default=None,
     help="The name of the workflow to list. Will display all versions",
 )
+@requires_login
 def get_wf(name: Union[str, None] = None):
     """List workflows."""
     crash_handler.message = "Unable to get workflows"
@@ -600,9 +602,9 @@ def get_wf(name: Union[str, None] = None):
         )
 
 
-@requires_login
 @main.command("preview")
 @click.argument("pkg_root", nargs=1, type=click.Path(exists=True, path_type=Path))
+@requires_login
 def preview(pkg_root: Path):
     """Creates a preview of your workflow interface."""
     crash_handler.message = f"Unable to preview inputs for {pkg_root}"
@@ -613,8 +615,8 @@ def preview(pkg_root: Path):
     preview(pkg_root)
 
 
-@requires_login
 @main.command("get-executions")
+@requires_login
 def get_executions():
     """Spawns an interactive terminal UI that shows all executions in a given workspace"""
 
@@ -630,7 +632,6 @@ LDATA COMMANDS
 """
 
 
-@requires_login
 @main.command("cp")
 @click.argument("src", shell_complete=cp_complete, nargs=-1)
 @click.argument("dest", shell_complete=cp_complete, nargs=1)
@@ -657,6 +658,7 @@ LDATA COMMANDS
     default=False,
     show_default=True,
 )
+@requires_login
 def cp(
     src: List[str],
     dest: str,
@@ -664,8 +666,7 @@ def cp(
     verbose: bool,
     no_glob: bool,
 ):
-    """Copy local files to LatchData and vice versa."""
-
+    """Copy files between Latch Data and local, or between two Latch Data locations. Behaves like `cp -R` in Unix. Directories are copied recursively. If any parents of dest do not exist, the copy will fail."""
     crash_handler.message = f"Unable to copy {src} to {dest}"
     crash_handler.pkg_root = str(Path.cwd())
 
@@ -680,7 +681,6 @@ def cp(
     )
 
 
-@requires_login
 @main.command("mv")
 @click.argument("src", shell_complete=remote_complete, nargs=1)
 @click.argument("dest", shell_complete=remote_complete, nargs=1)
@@ -692,6 +692,7 @@ def cp(
     default=False,
     show_default=True,
 )
+@requires_login
 def mv(src: str, dest: str, no_glob: bool):
     """Move remote files in LatchData."""
 
@@ -703,7 +704,6 @@ def mv(src: str, dest: str, no_glob: bool):
     move(src, dest, no_glob=no_glob)
 
 
-@requires_login
 @main.command("ls")
 @click.option(
     "--group-directories-first",
@@ -713,6 +713,7 @@ def mv(src: str, dest: str, no_glob: bool):
     default=False,
 )
 @click.argument("paths", nargs=-1, shell_complete=remote_complete)
+@requires_login
 def ls(paths: Tuple[str], group_directories_first: bool):
     """
     List the contents of a Latch Data directory
@@ -740,7 +741,6 @@ def ls(paths: Tuple[str], group_directories_first: bool):
             click.echo("")
 
 
-@requires_login
 @main.command("rmr")
 @click.argument("remote_path", nargs=1, type=str)
 @click.option(
@@ -767,6 +767,7 @@ def ls(paths: Tuple[str], group_directories_first: bool):
     default=False,
     show_default=True,
 )
+@requires_login
 def rmr(remote_path: str, yes: bool, no_glob: bool, verbose: bool):
     """Deletes a remote entity."""
     crash_handler.message = f"Unable to delete {remote_path}"
@@ -777,9 +778,9 @@ def rmr(remote_path: str, yes: bool, no_glob: bool, verbose: bool):
     rmr(remote_path, skip_confirmation=yes, no_glob=no_glob, verbose=verbose)
 
 
-@requires_login
 @main.command("mkdirp")
 @click.argument("remote_directory", nargs=1, type=str)
+@requires_login
 def mkdir(remote_directory: str):
     """Creates a new remote directory."""
     crash_handler.message = f"Unable to create directory {remote_directory}"
@@ -790,7 +791,6 @@ def mkdir(remote_directory: str):
     mkdirp(remote_directory)
 
 
-@requires_login
 @main.command("sync")
 @click.argument("srcs", nargs=-1)
 @click.argument("dst", nargs=1)
@@ -808,6 +808,7 @@ def mkdir(remote_directory: str):
     is_flag=True,
     default=False,
 )
+@requires_login
 def sync(srcs: List[str], dst: str, delete: bool, ignore_unsyncable: bool):
     """
     Update the contents of a remote directory with local data.
@@ -880,7 +881,6 @@ def test_data(ctx: click.Context):
         click.secho(f"{ctx.get_help()}")
 
 
-@requires_login
 @test_data.command("upload")
 @click.argument("src_path", nargs=1, type=click.Path(exists=True))
 @click.option(
@@ -891,6 +891,7 @@ def test_data(ctx: click.Context):
     type=bool,
     help="Automatically overwrite any files without asking for confirmation.",
 )
+@requires_login
 def test_data_upload(src_path: str, dont_confirm_overwrite: bool):
     """Upload test data object."""
 
@@ -903,9 +904,9 @@ def test_data_upload(src_path: str, dont_confirm_overwrite: bool):
     click.secho(f"Successfully uploaded to {s3_url}", fg="green")
 
 
-@requires_login
 @test_data.command("remove")
 @click.argument("object_url", nargs=1, type=str)
+@requires_login
 def test_data_remove(object_url: str):
     """Remove test data object."""
 
@@ -918,8 +919,8 @@ def test_data_remove(object_url: str):
     click.secho(f"Successfully removed {object_url}", fg="green")
 
 
-@requires_login
 @test_data.command("ls")
+@requires_login
 def test_data_ls():
     """List test data objects."""