+++ latch_cli/services/get_executions.py
@@ -10,8 +10,8 @@ from apscheduler.schedulers.background import BackgroundScheduler
 from latch_sdk_config.latch import config
 
 import latch_cli.menus as menus
+from latch.utils import account_id_from_token, current_workspace, retrieve_or_login
 from latch_cli.tinyrequests import post
-from latch_cli.utils import account_id_from_token, current_workspace, retrieve_or_login
 
 
 def get_executions():
@@ -50,12 +50,10 @@ def get_executions():
     ]
 
     for execution_data in sorted(data.values(), key=lambda x: -int(x["id"])):
-        options.append(
-            {
-                **execution_data,
-                "workflow_tagged": f'{execution_data["workflow_name"]}/{execution_data["workflow_version"]}',
-            }
-        )
+        options.append({
+            **execution_data,
+            "workflow_tagged": f'{execution_data["workflow_name"]}/{execution_data["workflow_version"]}',
+        })
 
     _all_executions(
         title="All Executions",
@@ -223,7 +221,8 @@ def _all_executions(
                 prev = (curr_selected, hor_index, term_width, term_height)
                 menus.clear_screen()
                 max_row_len = render(curr_selected, hor_index, term_width, term_height)
-    except KeyboardInterrupt: ...
+    except KeyboardInterrupt:
+        ...
     finally:
         menus.clear_screen()
         menus.reveal_cursor()
@@ -316,7 +315,8 @@ def _execution_dashboard(execution_data: Dict[str, str], workflow_graph: Dict):
                 menus.clear_screen()
                 prev = (curr_selected, term_width, term_height)
                 render(curr_selected, term_width, term_height)
-    except KeyboardInterrupt: ...
+    except KeyboardInterrupt:
+        ...
     finally:
         menus.clear_screen()
         menus.move_cursor((0, 0))
@@ -452,7 +452,8 @@ def _log_window(execution_data, fixed_workflow_graph: list, selected: int):
                 menus.clear_screen()
                 prev_term_dims = (vert_index, hor_index, term_width, term_height)
                 render(vert_index, hor_index, term_width, term_height)
-    except KeyboardInterrupt: ...
+    except KeyboardInterrupt:
+        ...
     finally:
         log_sched.shutdown()
         log_file.unlink(missing_ok=True)
@@ -513,7 +514,8 @@ def _abort_modal(execution_data):
             if prev_term_dims != (term_width, term_height):
                 prev_term_dims = (term_width, term_height)
                 render(term_width, term_height)
-    except KeyboardInterrupt: ...
+    except KeyboardInterrupt:
+        ...
     finally:
         menus.clear_screen()
         menus.move_cursor((0, 0))