+++ latch_cli/utils/__init__.py
@@ -13,13 +13,16 @@ from datetime import datetime, timedelta
 from enum import Enum
 from pathlib import Path
 from textwrap import dedent
-from typing import List
+from typing import List, TypedDict
 from urllib.parse import urljoin
 
 import click
+import gql
 import jwt
 from latch_sdk_config.user import user_config
+from latch_sdk_gql.execute import execute
 
+from latch.utils import account_id_from_token, current_workspace
 from latch_cli.click_utils import bold
 from latch_cli.constants import latch_constants
 from latch_cli.tinyrequests import get
@@ -81,30 +84,6 @@ def get_auth_header() -> str:
     return header
 
 
-def retrieve_or_login() -> str:
-    """Returns a valid JWT to access Latch, prompting a login flow if needed.
-
-    Returns:
-        A JWT
-    """
-
-    from latch_cli.services.login import login
-
-    token = user_config.token
-    if token == "":
-        token = login()
-    return token
-
-
-def current_workspace() -> str:
-    ws = user_config.workspace_id
-    if ws == "":
-        ws = account_id_from_token(retrieve_or_login())
-        user_config.update_workspace(ws, "Personal Workspace")
-
-    return ws
-
-
 def sub_from_jwt(token: str) -> str:
     """Extract a user sub (UUID) from a JWT minted by auth0.
 
@@ -124,39 +103,6 @@ def sub_from_jwt(token: str) -> str:
             " and is not a valid token."
         )
     return sub
-
-
-def account_id_from_token(token: str) -> str:
-    """Exchanges a valid JWT for a Latch account ID.
-
-    Latch account IDs are needed for any user-specific request, eg. register
-    workflows or copy files to Latch.
-
-    Args:
-        token: JWT
-
-    Returns:
-        A Latch account ID (UUID).
-    """
-
-    decoded_jwt = jwt.decode(token, options={"verify_signature": False})
-    try:
-        return decoded_jwt.get("id")
-    except KeyError as e:
-        raise ValueError("Your Latch access token is malformed") from e
-
-
-def _normalize_remote_path(remote_path: str):
-    if remote_path.startswith("latch://"):
-        remote_path = remote_path[len("latch://") :]
-    if (
-        not remote_path.startswith("/")
-        and not remote_path.startswith("shared")
-        and not remote_path.startswith("account")
-    ):
-        remote_path = f"/{remote_path}"
-
-    return remote_path
 
 
 def _si_number_strings(num):